Renamed assets to more descriptive names
