+++ kaleidoscopic-kings/assets/bonfire_icon.png
